--- .idea/.gitignore
@@ -1,8 +1,0 @@
-<<<<<<< Updated upstream
-
-# Datasource local storage ignored files
-/dataSources/
-=======
->>>>>>> Stashed changes
-# Datasource local storage ignored files
-/dataSources.local.xml
