+++ src/com/ss/lms/Main.java
@@ -1,6 +1,8 @@
 package com.ss.lms;
 
 import com.ss.dao.*;
+import com.ss.lms.model.Borrower;
+import com.ss.lms.services.AdminServices;
 import com.ss.tools.*;
 
 import java.sql.SQLException;
@@ -10,16 +12,16 @@ import java.util.Scanner;
 
 public class Main {
 	
-	
+
 	
 	static UI ui = new UI();
 	static BorrowerView borrowerView = new BorrowerView();
-	
+
 	public static Scanner userInput = new Scanner(System.in);
 	
-	public static void main(String[] args) throws SQLException {
-		
-		
+	public static void main(String[] args)  {
+
+
 		char input = ' ';
 
 		ui.mainMenuTop();
