+++ src/com/ss/lms/model/Publisher.java
@@ -1,4 +1,4 @@
-package com.ss.lms.Model;
+package com.ss.lms.model;
 
 public class Publisher {
     private Integer publisherId;
