+++ src/com/ss/lms/secret/Url.java
@@ -4,7 +4,7 @@ public class Url {
      public String getUrl()
     {
 
-        ServerCred mySecret = new ServerCred();
+        ServerCredential mySecret = new ServerCredential();
         return ("jdbc:mysql://localhost:3306/library?" + "user=" +
                 mySecret.getServerUser() + "&password="+
                 mySecret.getServerPWord()+"&serverTimezone=EST5EDT");
